requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Async and WebSocket support
websockets==12.0
websocket-client==1.6.4

# Task queue and messaging
celery==5.3.4
redis==5.0.1
kafka-python==2.0.2
asyncio-mqtt==0.16.1
aiokafka==0.10.0
confluent-kafka==2.3.0
avro-python3==1.10.2

# Audio processing and ML
numpy==1.24.3
librosa==0.10.1
speechrecognition==3.10.0
# pyaudio==0.2.11  # Commented out due to build issues on Windows
torch==2.1.1
transformers==4.35.2

# Environment and configuration
python-dotenv==1.0.0
openai==1.101.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CI/CD and Code Quality Tools
flake8==7.0.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
bandit==1.7.5
safety==3.6.1
pre-commit==3.6.0

# Security packages
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-limiter==0.1.6
cryptography==41.0.7
bcrypt==4.1.2

# Monitoring and ML tracking packages
mlflow==2.8.1
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
psutil>=6.1.0
wandb==0.16.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Real-time processing and gRPC packages
grpcio==1.59.3
grpcio-tools==1.59.3
grpcio-status==1.59.3
protobuf==4.25.1
asyncio-mqtt==0.16.1
aiokafka==0.10.0
confluent-kafka==2.3.0
avro-python3==1.10.2

# Additional dependencies
requests==2.31.0
websocket-client==1.6.4